namespace restructure, resolves #6
all tests pass, removed lots of shared
modules, kept all of mine/ones that I've
rewritten

still lots of changes need to be made
