--- my/config.py
@@ -1,2 +1,0 @@
-# TODO ok, this thing should trigger .cfg import presumably??
-from .core import init
