--- my/core/__init__.py
@@ -1,9 +1,0 @@
-# this file only keeps the most common & critical types/utility functions
-from .common import PathIsh, Paths, Json
-from .common import get_files
-from .common import LazyLogger
-from .common import warn_if_empty
-from .common import stat, Stats
-from .common import __NOT_HPI_MODULE__
-
-from .cfg import make_config
