+++ my/games/steam.py
@@ -32,7 +32,7 @@ from itertools import chain
 
 from ..core import get_files, Stats
 from ..core.error import Res
-from ..core.time import parse_datetime_sec
+from ..utils.time import parse_datetime_sec
 
 
 def inputs() -> Sequence[Path]:
