+++ my/ttt.py
@@ -33,8 +33,7 @@ from typing import Sequence
 
 from .core import get_files, warn_if_empty, Stats
 from .core.common import listify
-from .core.time import parse_datetime_sec
-from .core.file import filter_subfile_matches
+from .utils.time import parse_datetime_sec
 
 
 @listify
@@ -69,7 +68,7 @@ Results = Iterator[Entry]
 
 
 def history(from_paths=inputs) -> Results:
-    yield from _merge_histories(*map(_parse_file, filter_subfile_matches(from_paths())))
+    yield from _merge_histories(*map(_parse_file, from_paths()))
 
 
 @warn_if_empty
