+++ my/window_watcher.py
@@ -39,8 +39,7 @@ from itertools import chain
 
 from .core import get_files, warn_if_empty, Stats
 from .core.common import listify
-from .core.time import parse_datetime_sec
-from .core.file import filter_subfile_matches
+from .utils.time import parse_datetime_sec
 
 
 @listify
@@ -85,7 +84,7 @@ def history(from_paths=inputs) -> Results:
     yield from _construct_stream(
         filter(
             _is_unlikely,
-            _merge_histories(*map(_parse_file, filter_subfile_matches(from_paths()))),
+            _merge_histories(*map(_parse_file, from_paths())),
         )
     )
