+++ my/zsh.py
@@ -39,8 +39,7 @@ from typing import Sequence
 
 from .core import get_files, warn_if_empty, Stats
 from .core.common import listify
-from .core.time import parse_datetime_sec
-from .core.file import filter_subfile_matches
+from .utils.time import parse_datetime_sec
 
 
 @listify
@@ -84,7 +83,7 @@ Results = Iterator[Entry]
 
 
 def history(from_paths=inputs) -> Results:
-    yield from _merge_histories(*map(_parse_file, filter_subfile_matches(from_paths())))
+    yield from _merge_histories(*map(_parse_file, from_paths()))
 
 
 def stats() -> Stats:
