+++ pytest.ini
@@ -2,9 +2,10 @@
 # for now, running with with_my pytest tests/reddit.py
 # discover __init__,py as well
 # TODO also importing too much with it...
-#python_files = reddit.py
+# reddit.py is temporary, till I merge pushshift comments
 addopts =
   --ignore=tests/pdfs
+  --ignore=tests/test_reddit.py
   --random-order
   --verbose
